Remove default of setting user to root
ECS default behavior is to have this left unset, allowing the container
to run as the user the container was built as. Defaulting the runtime is
a security gap as it can cause containers that were built to run as
other users to suddenly start running as root when deployed with this
module.

Note, this could break existing container environments running in the
wild if they were inadvertently taking advantage of the root access.