Force user to "0" when firelens is set (#97)
* Force user to "0" when firelens is set

Recently I started experiencing apply churn in my runs. I eventually traced it down to a behavior with Firelens enabled task.

If one submits a firelens-enabled task definition with `user=null`, AWS helpful replaces the value with `"0"`. This causes Terraform to repeatedly try to fix the discrepancy. This change aims to address this edge-case behavior which I found to be unintuitive.

                  ~ user                   = "0" -> null

Additionally, attempting to set the `user` to any value _other_ than `"0"` (or null) is an error:

                  ~ user                   = "0" -> "root"
    ...
    Error: ClientException: If 'user' field is specified on firelens container, then 'UID' has to be '0'.

Which simplifies our ternary, so that's nice. It would have preferred to deal with this with a [custom validation rule](https://www.terraform.io/docs/configuration/variables.html#custom-validation-rules), unfortunately that doesn't support checking multiple inputs so it's not viable here.

* Executed 'terraform fmt'

* Updated README.md

Co-authored-by: actions-bot <58130806+actions-bot@users.noreply.github.com>